# src/users/dto/update-user.dto.js
import Joi from 'joi';

class UpdateUserDto {
  constructor(user = {}) {
    this.email = user.email;
    this.password = user.password;
  }

  static swaggerSchema() {
    return {
      email: 'updated@example.com',
      password: 'newpassword123',
    };
  }

  static userSchema = Joi.object({
    email: Joi.string().email().required(),
    password: Joi.string().min(6).required(),
  });

  validate() {
    const { error } = UpdateUserDto.userSchema.validate(
      {
        email: this.email,
        password: this.password,
      },
      { abortEarly: false },
    );

    return {
      valid: !error,
      errors: error ? error.details.map((detail) => detail.message) : [],
    };
  }
}

export { UpdateUserDto };